Make the tests mostly pass in Python 3.6
There is an issue with the deep dependency, which isn't Python 3-compatible.

+++ complaint_search/tests/test_stream_content.py
@@ -1,4 +1,5 @@
 from django.test import TestCase
+
 from complaint_search.stream_content import (
     StreamCSVContent,
     StreamJSONContent,
@@ -37,7 +38,7 @@ class StreamJSONContentTests(TestCase):
 
         # pretend this is broken up randomly every 20 chars
         self.content_list = [self.content[(i * 20):(i * 20 + 20)]
-                             for i in range(len(self.content) / 20 + 1)]
+                             for i in range(int(len(self.content) / 20 + 1))]
 
     def test_iter(self):
         sc = StreamJSONContent(iter(self.content_list))
@@ -46,7 +47,7 @@ class StreamJSONContentTests(TestCase):
     def test_next_complete(self):
         for size in range(1, 1024):
             content_list = [self.content[(i * size):(i * size + size)]
-                            for i in range(len(self.content) / size + 1)]
+                            for i in range(int(len(self.content) / size + 1))]
             sc = StreamJSONContent(iter(content_list))
             result = ""
             for json_in_progress in sc:
